make client init and unary call work
Apart from exceptions below, this code generates
image_annotator_client.go byte-for-byte identical with the one
in production.

- The foos placeholders
- LRO methods and types aren't implemented yet
- Imports aren't implemented yet

There's also other "turn-off logic" missing. For example,
we unconditionally include LRO client, while it should only be
included in APIs with LRO.
This is OK for now. Soon we'll generate a client for API without
LRO and match it against the one in production;
we can deal with it then.

I'm also experimenting with using fewer abstractions, so I currently
eschew templates in favor of print statements. I'm not married to
the idea though.

Change-Id: I34f467e6b0dc473808484a8db00fc23070f66e9a
Reviewed-on: https://gapi-review.git.corp.google.com/6550
Reviewed-by: Eva Ogbe <eoogbe@google.com>