generate LRO types
The LRO response and metadata types are placeholders for now,
since we need additional information from protobuf annotations.

It is rather awkward to read the annotations right now,
since we need to build the definition of the annotations.
My current plan is to hold off on reading the annotations
until their definitions are in googleapis master branch.
It should be an easy change to make.

Change-Id: I4d0476f1b8deb8d208c78f0c0e8ad55e1faeb282
Reviewed-on: https://gapi-review.git.corp.google.com/6591
Reviewed-by: Michael Darakananda <pongad@google.com>