generate streaming methods
None of the currently proto-annotated cloud APIs use server or client
streaming right now, so I haven't gotten a chance to baseline-test this
against production. Still, I believe problems (if any) will be small
and easily fixed in a followup, and we'll likely catch them when we
test with Showcase [1].

[1] Showcase is an API we created to exercise various functionalities
of the generator. It has an in-memory implementation we can test
against.