Have gengapic protoc plugin invoke samplegen (#218)
Move most of cmd/gen-go-sample to intenal/gensample, but leave main.go still in cmd/gen-go-sample. This makes sure that samplegen can continue to run as a standalone program, and gapic-gen and invoke samplegen

Add --sample, --gapic-config and --sample-only flag to docker entry point, and pass them along to gensample and gengapic.

When --sample-only is present, toggle off gengapic and gapic_validator_out, so that we can generate samples for APIs that are not yet properly annotated.
gensample will consume gapic config and sample config.