chore: wrap gazelle update-repo in make (#451)
This wraps the `gazelle` `update-repos` command in Make target to simplify updating the Bazel repositories with the Go modules.

`gazelle` will only generate `go_repository` rules to load Go module dependencies, but we like to wrap our dependencies with a `_maybe` macro, in case another section earlier in a WORKSPACE has already imported the dependency. So, I've defined a `go_repository` macro that invokes `gazelle`'s `go_repository` with our `_maybe`. `gazelle` `update-repos` continues to generate `go_repository` rules from our `go.mod`. 

We also have to do a little find-remove because `gazelle` generates a load statement for `go_repository` even though we have already loaded it with an alias (`gazelle_go_repository = "go_repository"`).

Finally, the last hack, the `go.sum` declares `cloud.google.com/go` as a transitive dependency brought in at some point in the tree, thus making `gazelle` include it as a `go_repository`. The `gapic-generator-go` does not depend on this module at all and it shouldn't be imported as a build dependency of the generator. So, we allow `gazelle` to include the dep, but we use our `go_repository` wrapper to explicitly ignore it. This is easier than more find-remove scripting.