feat: add interface abstraction primitives
Nomenclature within this commit:
* The current generated client is referred to as a 'transport'.
  Similarly, the WIP rest+json client will also be a 'transport'.
* The new internal-only interface is the 'transport interface'.
* The new client type whose only responsibility is to embed a
transport interface is a 'new-style client'.

Two new types are defined: an interface intended to abstract out the
mechanism of network communication and serialization (the transport interface),
and a wrapper struct that embeds the interface.

The transport interface cannot be user visible because methods will be
added to it if the API defines new methods. This is a non-breaking
change in the API but is technically a breaking change to the user; if
they defined a type that previously satisfied the contract of the
interface, the type would no longer satisfy the interface.

The workaround is to make the transport interface not user visible and
to embed the interface in a struct. The transports and transport
interface are generated concurrently, so there is no risk of breaking
the interface contract. Adding methods to the new-style client is a
non-breaking change becasuse it is a struct.
This is similar to the PIMPL idiom in C++.

TODO:
* Add factory function(s) for the new-style client to take an injected
transport + auth.
* Tweak documentation, samples, and integration tests to prefer
new-style clients instead of directly using transports.
* Write out rest+json transport.

Includes moderately substantial refactors as implementation details.