feat: Integration test:  Showcase compliance suite (#625)
This tests the Showcase compliance suite against a running Showcase server. The client-side checks are merely that the returned data equals the sent data. However, when the requests are made over REST, the Showcase server will check various aspects of the REST transcoding, including optional field presence, headers, etc., and informatively fail the call if any of the checks fail.

**NOTE**: At the moment gapic-generator-go does not yet expose a factory function for REST clients. Using that when available remains a TODO. For the time being, we're using a gRPC client, which will trivially pass the test.