feat: internal methods for determining path and query parameters
When using REST transports, request message fields can be separated
into "path" parameters and "query" parameters.
These parameters are used to modify the the request URL in different
ways, so it is important to determine which set a given field belongs
to.

This PR is a scaffolding PR that adds internal generator methods and
associated tests to partition request message fields into path and
query params.

These fields should have no intersection.