gapic: wrap diregapic custom operations (#722)
In `diregapic` mode, a new plugin flag, the generator will look for a type named `Operation` in the proto-package-to-generate, and wherever it is used as a return type for an RPC, it will be replaced by an `Operation` type generated in the gapic's package that wraps the proto-defined `Operation`.  This will allow us to implement long running operation helpers in the future by extending the generated `Operation` wrapper with new functionality. Without this, we'd most likely need to take a breaking change to wrap the proto-defined `Operation` and extend it with helpers.

Companion PR for `compute`: https://github.com/googleapis/google-cloud-go/pull/4511