Add an error check after generated CLI functions call service methods.
A generated CLI handler was crashing when it called an API method
that returned an error. The method was expected to return an LRO,
and the handler crashed when it called "resp.Name()" for a nil resp.
"resp, err :=" was set with the call to the API method and there
was no check of the returned err value. Adding
	if err != nil {return err}
eliminated the crash.