feat(gengapic): rest-numeric-enums option enables enum-encoding sys param (#1022)
For REGAPIC API calls, when the `rest-numeric-enums` plugin option is provided, include the new system parameter variant of `$alt=json` that enables proto-json encoding of enums in server responses as integers - `enum-encoding=int`.  The default is false. Includes the bazel attribute as well.

Also updates the README with option documentation.

Requested in http://b/232457394.