fix(internal/gengapic): add workaround for operation collision
We have a service that is not yet released that uses the same RPC
name across two services in the same proto package. Our Operation
wrapper identifiers are not scoped to a service so this causes
issues when generated a client that does this. For now this is for
one client, which may change before GA, so a hacky solution is fine.
But we should keep this in the back of our minds should we ever v2
our libraries this is something we would want to avoid.