chore(internal/gengapic): refactor g.init to newGenerator
At the moment, the `g.init` modifies a type `generator`. In many cases,
`g` is empty and populated by `g.init`. It is generally more idiomatic
to use a `new<thing>` function to construct a new instance of a type, so
g.init is replaced with a `newGenerator` function to follow that
pattern.