feat: add proto runtime version in header (#1624)
We are adding the version of google.golang.org/protobuf that we
are using in the x-goog header. The library currently only exposes
their minor version. If they ever take a v2 we will need to update
this code accordingly and/or ask them to expose the major version as
well.

Internal Bug: 415822167

* add comment